requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11
djangorestframework==3.6.2
heroku==0.1.4
idna==2.8
Pillow==5.4.1
python-dateutil==1.5
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
